backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# ML and embeddings (from your existing project)
sentence-transformers==2.2.2
chromadb==0.4.18
numpy==1.24.3

# Astronomical calculations (from your existing project)
pyephem==4.1.5
swisseph==2.10.03.1

# Utility
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Logging and monitoring
python-json-logger==2.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1